+++ __tests__/ImageHelperTest.php
@@ -14,26 +14,40 @@ final class ImageHelperTest extends TestCase {
     FileHelper::copyDirectory(INPUT_BAK_DIR, INPUT_DIR);
   }
 
-  public function testWriteFirstFrameOfGifInSeparateFile(): void {
-    $inputPath = INPUT_DIR . '/sample.gif';
-    $outputPath = OUTPUT_DIR .  '/write-first-frame-of-gif-in-separate-file.gif';
-    ImageHelper::extractFirstFrameOfGif($inputPath, $outputPath);
-    $this->assertSame(ImageHelper::getNumberOfGifFrames($outputPath), 1);
+  // public function testWriteFirstFrameOfGifInSeparateFile(): void {
+  //   $inputPath = INPUT_DIR . '/sample.gif';
+  //   $outputPath = OUTPUT_DIR .  '/write-first-frame-of-gif-in-separate-file.gif';
+  //   ImageHelper::extractFirstFrameOfGif($inputPath, $outputPath);
+  //   $this->assertSame(ImageHelper::getNumberOfGifFrames($outputPath), 1);
+  // }
+
+  // public function testWriteFirstFrameOfGifInOriginalFile(): void {
+  //   $inputPath = INPUT_DIR . '/sample.gif';
+  //   ImageHelper::extractFirstFrameOfGif($inputPath);
+  //   $this->assertSame(ImageHelper::getNumberOfGifFrames($inputPath), 1);
+  // }
+
+  // public function testGetNumberOfGifFrames(): void {
+  //   $inputPath = INPUT_DIR . '/sample_2.gif';
+  //   $this->assertSame(ImageHelper::getNumberOfGifFrames($inputPath), 19);
+  // }
+
+  // public function testGetNumberOfFramesInGifWithNoAnimation(): void {
+  //   $inputPath = INPUT_DIR . '/non-animated.gif';
+  //   $this->assertSame(ImageHelper::getNumberOfGifFrames($inputPath), 1);
+  // }
+
+  public function testConvertAllPagesOfPdfToJpg(): void {
+    $inputPath = INPUT_DIR . '/sample.pdf';
+    $outputPath = OUTPUT_DIR .  '/pdf.jpg';
+    ImageHelper::pdf2Image($inputPath, $outputPath);
+    $this->assertSame(true, true);
   }
 
-  public function testWriteFirstFrameOfGifInOriginalFile(): void {
-    $inputPath = INPUT_DIR . '/sample.gif';
-    ImageHelper::extractFirstFrameOfGif($inputPath);
-    $this->assertSame(ImageHelper::getNumberOfGifFrames($inputPath), 1);
-  }
-
-  public function testGetNumberOfGifFrames(): void {
-    $inputPath = INPUT_DIR . '/sample_2.gif';
-    $this->assertSame(ImageHelper::getNumberOfGifFrames($inputPath), 19);
-  }
-
-  public function testGetNumberOfFramesInGifWithNoAnimation(): void {
-    $inputPath = INPUT_DIR . '/non-animated.gif';
-    $this->assertSame(ImageHelper::getNumberOfGifFrames($inputPath), 1);
+  public function testConvertOnlyOnePageOfPdfToJpg(): void {
+    $inputPath = INPUT_DIR . '/sample.pdf';
+    $outputPath = OUTPUT_DIR .  '/pdf.jpg';
+    ImageHelper::pdf2Image($inputPath, $outputPath, ['pageNumber' => 0]);
+    $this->assertSame(true, true);
   }
 }